pyproject.toml
[project]
name = "36-webapp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.0",
    "bcrypt==5.0.0",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "coverage==7.13.1",
    "cryptography==46.0.3",
    "django==5.2.9",
    "django-jalali==7.4.0",
    "djangorestframework==3.16.1",
    "exceptiongroup==1.3.1",
    "factory-boy==3.3.3",
    "faker==40.1.0",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jalali-core==1.0.0",
    "jdatetime==5.2.0",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "packaging==25.0",
    "pillow==12.1.0",
    "pluggy==1.6.0",
    "psutil==7.2.1",
    "pycparser==2.23",
    "pygments==2.19.2",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "pytest-django==4.11.1",
    "requests==2.32.5",
    "rich==14.2.0",
    "sqlparse==0.5.5",
    "tomli==2.3.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "urllib3==2.6.2",
]
